Make dataselectormodal exportable
